Created LoadTileSet function
As a start to me trying to increase the map loading speed, I first created a function "LoadTileSet" which creates a vector that stores smart pointers to tile objects that represent each tile in the tileset. The pointers can be used when referencing these tiles instead of creating a new tile object for each tile in the map.