Added in loading screen on game start
With the current way level data is loaded, there is a significant delay at startup of the application to get in the game since all the levels' tile data has to be allocated at this time. To make the wait more interesting, I have added a loading screen using title image created by the team with a few changes which is shown during the load time.